Restructure plugin files to clean up and structurize
