+++ assets/js/color_picker.js
