+++ assets/js/notice-update.js
