+++ includes/widgets/kbe-widget-category.php
