+++ includes/widgets/kbe-widget-search.php
